Moved LinkStrategies and SourceSnapshot into dedicated source files.
